Fix issue with data migration always failing after it was successful

+++ app/src/main/java/paulscode/android/mupen64plusae/task/ExtractAssetsOrCleanupTask.java
@@ -261,6 +261,10 @@ public class ExtractAssetsOrCleanupTask extends AsyncTask<Void, String, List<Ext
 
     private boolean createBackupAndMove(String srcPath, String dstPath)
     {
+        if (!new File(srcPath).exists()) {
+            return true;
+        }
+        
         boolean backupSuccess = FileUtil.copyFile(new File(srcPath), new File(srcPath + ".bak"));
         boolean copySuccess = FileUtil.copyFile(new File(srcPath), new File(dstPath));
